+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### Ignition Rendering 2.X.X (2019-XX-XX)
 
+1. Ogre2: Fix shadows by dynamically generating shadow maps
+    * [Pull request 222](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/222)
+    * [Pull request 223](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/223)
+
 ### Ignition Rendering 2.2.0 (2019-11-20)
 
 1. Backport setting cast shadows